Improve flexibility of authors, affiliations input
Specification of authors and their affiliations must be as simple and
intuitive as possible. Processing of the `author` and `institute` meta
fields is hence improved. The number of possible ways to input this data
is increased, improving flexibility and convenience for authors.

Meta data for authors and affiliations are uniformly treated as named
objects. Names can be given as a string, as via the key of a
single-element object, or explicitly as a `name` attribute of an
object. The following are hence equivalent:

    author:
      - John Doe:
          institute: Agile Science Society

    author:
      - name: John Doe
        institute: Agile Science Society

    author:
      - name: John Doe
        institute:
          - name: Agile Science Society

Institutes may be given separately. If `institute` is an object, the elements
therein can be referred to using their key.

    author:
      - John Doe:
          institute: agsci
    institute:
      agsci: Agile Science Society

If `institute` is an array, numeric indices can be used instead.

    author:
      - John Doe:
          institute: 1
    institute:
      - Agile Science Society

Scripts of the publishing pipeline are adapted accordingly.