Add rudimentary support for CiTO ontologies
The type of a citations can be specified using `@CITO_PROPERTY:KEY`,
where `CITO_PROPERTY` is a supported Citation Ontology Typing property,
and `KEY` is the usual citation key. A general default
property (`cites`) is used if no CiTO property is found.

This is currently only included in the JSON-LD writer.  The properties
are stripped in all other exports.